Disabled Node24 download for linux-armv7l

+++ src/Misc/externals.sh
@@ -6,6 +6,7 @@ L1_MODE=$4
 
 INCLUDE_NODE6=${INCLUDE_NODE6:-true}
 INCLUDE_NODE10=${INCLUDE_NODE10:-true}
+INCLUDE_NODE20=${INCLUDE_NODE20:-true}
 INCLUDE_NODE24=${INCLUDE_NODE24:-true}
 
 CONTAINER_URL=https://vstsagenttools.blob.core.windows.net/tools
@@ -21,6 +22,10 @@ fi
 if [[ "$PACKAGERUNTIME" == "win-arm64" ]]; then
     INCLUDE_NODE6=false
     INCLUDE_NODE10=false;
+fi
+
+if [[ "$PACKAGERUNTIME" == "linux-arm" ]]; then
+    INCLUDE_NODE24=false
 fi
 
 NODE_VERSION="6.17.1"
@@ -298,7 +303,9 @@ else
         fi
         acquireExternalTool "${NODE_URL}/v${NODE16_VERSION}/node-v${NODE16_VERSION}-${ARCH}.tar.gz" node16 fix_nested_dir
         acquireExternalTool "${NODE_URL}/v${NODE20_VERSION}/node-v${NODE20_VERSION}-${ARCH}.tar.gz" node20_1 fix_nested_dir
-        acquireExternalTool "${NODE_URL}/v${NODE24_VERSION}/node-v${NODE24_VERSION}-${ARCH}.tar.gz" node24 fix_nested_dir
+        if [[ "$INCLUDE_NODE24" == "true" ]]; then
+            acquireExternalTool "${NODE_URL}/v${NODE24_VERSION}/node-v${NODE24_VERSION}-${ARCH}.tar.gz" node24 fix_nested_dir
+        fi
     fi
     # remove `npm`, `npx`, `corepack`, and related `node_modules` from the `externals/node*` agent directory
     # they are installed along with node, but agent does not use them
@@ -320,10 +327,12 @@ else
     rm "$LAYOUT_DIR/externals/node20_1/bin/npx"
     rm "$LAYOUT_DIR/externals/node20_1/bin/corepack"
 
-    rm -rf "$LAYOUT_DIR/externals/node24/lib"
-    rm "$LAYOUT_DIR/externals/node24/bin/npm"
-    rm "$LAYOUT_DIR/externals/node24/bin/npx"
-    rm "$LAYOUT_DIR/externals/node24/bin/corepack"
+    if [[ "$INCLUDE_NODE24" == "true" ]]; then
+        rm -rf "$LAYOUT_DIR/externals/node24/lib"
+        rm "$LAYOUT_DIR/externals/node24/bin/npm"
+        rm "$LAYOUT_DIR/externals/node24/bin/npx"
+        rm "$LAYOUT_DIR/externals/node24/bin/corepack"
+    fi
 fi
 
 if [[ "$L1_MODE" != "" || "$PRECACHE" != "" ]]; then
